+++ tests/Fixtures/CustomerImport.php
@@ -9,13 +9,14 @@ use Ivanfuhr\Ingestor\Contract\Context;
 use Ivanfuhr\Ingestor\Contract\Definition;
 use Ivanfuhr\Ingestor\Dataset\Dataset;
 use Ivanfuhr\Ingestor\Row\Row;
+use Ivanfuhr\Ingestor\Schema\DatasetBuilder;
 use Ivanfuhr\Ingestor\Schema\Schema;
 use Ivanfuhr\Ingestor\Stage\EmptyStage;
 use Ivanfuhr\Ingestor\Stage\PrefilledStage;
 
 final class CustomerImport implements Definition
 {
-    public function schema(): Schema
+    public function schema(): Schema|DatasetBuilder
     {
         return Schema::make()
             ->dataset('customers')