+++ tests/Fixtures/SimpleCustomerImport.php
@@ -8,12 +8,13 @@ use Ivanfuhr\Ingestor\Contract\Context;
 use Ivanfuhr\Ingestor\Contract\Definition;
 use Ivanfuhr\Ingestor\Dataset\Dataset;
 use Ivanfuhr\Ingestor\Row\Row;
+use Ivanfuhr\Ingestor\Schema\DatasetBuilder;
 use Ivanfuhr\Ingestor\Schema\Schema;
 use Ivanfuhr\Ingestor\Stage\EmptyStage;
 
 final class SimpleCustomerImport implements Definition
 {
-    public function schema(): Schema
+    public function schema(): Schema|DatasetBuilder
     {
         return Schema::make()
             ->dataset('customers')